requirements.txt
psycopg2-binary==2.9.3
djangorestframework==3.15.1
python-decouple==3.8
Django==5.0.2
django-phonenumber-field==7.3.0
